ser/requirements.txt
flask==2.0.1
flask-cors==3.0.10
gunicorn==20.1.0
werkzeug==2.0.3
act-workflow==4.8.243